requirements.txt
Django>=2.2.3,<2.3.0
djangorestframework>=3.10.1,<3.11.0

psycopg2-binary>=2.8.3,<2.9.0

flake8>=3.7.8,<3.8.0

matplotlib>=3.1.0,<3.2.0
pandas>=1.0.1,<1.1.0
numpy>=1.18.1,<1.20.0
Pillow>=7.0.0,<7.1.0
django-cors-headers
gunicorn>=20.0.4,<20.1.0
torch>=1.0.0,<=1.11.0
torchvision>=0.2.2,<=0.11.3